361.2

+++ src/PixelFarm/PixelFarm.MiniAgg/04_Scanline/4.1_ScanlineRasterizer.cs
@@ -186,7 +186,7 @@ namespace PixelFarm.Agg
         {
             for (int i = AA_SCALE - 1; i >= 0; --i)
             {
-                m_gammaLut[i] = (int)AggMath.uround(
+                m_gammaLut[i] = AggMath.uround(
                     gamma_function.GetGamma((float)(i) / AA_MASK) * AA_MASK);
             }
         }
